makefile: replace yui-compressor with tenser
`yui-compressor` is still available from the Debian packages; but it's no longer
maintained, and isn't useful for ES6. Replace it with `tenser` off of NPM, and
rebuild.

Surprisingly, this necessitates changing the outer IIFE to use an arrow-
function. `tenser` transforms `(function(){...})();` into `!function(){...}();`
to save a character, but this also changes the return value from `undefined` to
`true` -- and Firefox cares about the return value of bookmarklets!

(If you're wondering why the original code used `function`: arrow functions were
available in Firefox in 2013, but didn't make it into Chrome, Edge, and Opera
until late 2015, nor into Safari until 2016.)